+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Changed
  - MockLightingProcessor now implements IMockDevice
  - MockShadeDevice now implement IMockDevice
+ - Using new logging context
 
 ## [5.0.1] - 2020-04-05
 ### Changed